Don't copy images if they already exist

+++ colmap2nerf.py
@@ -144,7 +144,8 @@ output_image_dir.mkdir(exist_ok=True, parents=True)
 
 for image_id in scene_manager.image_ids:
     img_path = scene_manager.path_to_image(image_id)
-    shutil.copy(img_path, output_image_dir)
+    if not img_path.exists():
+        shutil.copy(img_path, output_image_dir)
 
 print(f'Saved images to {output_image_dir}')
